finish JWT routes

+++ server/src/routes/memories.ts
@@ -4,12 +4,17 @@ import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
 
-  app.get('/test', async (request, reply) => {
-    return 'Olá'
+  //Vefiricar para todas as rotas, se o usuário está logado
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
   })
 
   app.get('/memories', async (request, reply) => {
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       }
@@ -38,6 +43,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
       }
     })
 
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     return memory
 
   })
@@ -57,7 +66,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'c387ce71-b128-4f11-bc35-4ce672625626',
+        userId: request.user.sub,
       }
     })
 
@@ -81,7 +90,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = createMemoryBody.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUnique({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -102,6 +121,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
 
     const { id } = memoryParamsSchema.parse(request.params);
+
+    let memory = await prisma.memory.findUnique({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
